UROP 2022

+++ CPFA/build.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
-#if [ "$1" = "clean" ]; then
+if [ "$1" = "clean" ]; then
     echo "Deleting and recreating the build directory "
     rm -rf build
     mkdir build
-#fi
+fi
 
 cd build
 export PKG_CONFIG_PATH=/opt/local/argos3/2.8.12.2/gcc/5.4.0/lib/pkgconfig
